export/pyproject.toml
[project]
name = "export"
version = "0.1.0"
description = "Export service for fine-tuned Gemma models"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.116.1",
    "google-cloud-firestore>=2.21.0",
    "google-cloud-storage>=3.3.0",
    "peft>=0.17.1",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
    "python-multipart>=0.0.20",
    "transformers>=4.56.0",
    "unsloth>=2025.8.10",
    "uvicorn>=0.35.0",
]
